Revert "Unittest/restaurant"

+++ src/main/java/org/example/easytable/member/controller/MemberController.java
@@ -33,7 +33,7 @@ public class MemberController {
 
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+	public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
 		memberService.deleteUser(id);
 		return ResponseEntity.noContent().build();
 	}
